+++ packages/react-cosmos-shared2/src/util/index.ts
@@ -1,39 +1,4 @@
-import { findIndex } from 'lodash';
-
 export type StateUpdater<T> = (prevState: T) => T;
 
-export function updateItem<T>(
-  items: Readonly<T[]>,
-  item: T,
-  update: Partial<T>
-): T[] {
-  const index = items.indexOf(item);
-  return [
-    ...items.slice(0, index),
-    { ...item, ...update },
-    ...items.slice(index + 1)
-  ];
-}
-
-export function replaceOrAddItem<T>(
-  items: Readonly<T[]>,
-  matcher: (item: T) => boolean,
-  item: T
-): T[] {
-  const index = findIndex(items, matcher);
-  return index !== -1
-    ? [...items.slice(0, index), item, ...items.slice(index + 1)]
-    : [...items, item];
-}
-
-export function removeItemMatch<T>(
-  items: Readonly<T[]>,
-  matcher: (item: T) => boolean
-): T[] {
-  const index = findIndex(items, matcher);
-  return index === -1
-    ? [...items]
-    : [...items.slice(0, index), ...items.slice(index + 1)];
-}
-
+export { updateItem, replaceOrAddItem, removeItemMatch } from './array';
 export { uuid } from './uuid';
